Add encryption to background.ts
The user inputs their text in the popup (completely isolated from the
website), and the plaintext is sent to the background (also isolated
from the website) to be encrypted. Once we encrypt the text we send to
the content script to embed the ciphertext into the page. No plaintext
ever touches the website DOM.