Move decryption from popup to iframe
On OSX, decryption was not working because the element picker would act
up if the popup was open at the same time. Similar behavior was shown
with uBlock's element picker. We take uBlock's idea and embed an iframe
inside the target page to display decrypted content. Further, we assume
the page is malicious, and use web_accessible_resources as the iframe
source. Hostile javascript running on the page will not be able to
access decrypted content due to the same-origin policy.