Properly dismiss element picker
* Fix up event listeners while element picker is active.

The event.stopPropagation and event.preventDefault let the user dismiss
the picker without the page intercepting the mouse click or escape key.

* Add a callback function when the picker is dismissed

Use the message passing API to send an event to the isolated javascript,
telling it to dismiss the decrypted text popup.

* Move picker creation to chrome message listener

No hacky timeouts to insert the decryption iframe into the page. Just
add it when the user starts the picker.